+++ src/endpoints/pages.py
@@ -1,17 +1,16 @@
 # -*- coding: utf-8 -*-
 
-import datetime
-import time
+
 from fastapi import APIRouter, Request
 from fastapi.responses import RedirectResponse
+from httpx import AsyncClient
 from loguru import logger
 from sqlalchemy import Select
-from httpx import AsyncClient
+
 from ..db_tables import InterestingThings, Posts
+from ..functions.interesting_api_calls import get_public_debt
 from ..resources import db_ops, templates
 from ..settings import settings
-from unsync import unsync
-from tqdm import tqdm
 
 router = APIRouter()
 client = AsyncClient()
@@ -105,13 +104,13 @@ async def public_debt(request: Request):
     last_year = None
     for d in debt_list:
         year_hold = d["tot_pub_debt_out_amt"]
-        
+
         if d["debt_held_public_amt"] != 'null':
             d["debt_held_public_amt"] = "{:,.2f}".format(float(d["debt_held_public_amt"]))
-        
+
         if d["intragov_hold_amt"] != 'null':
             d["intragov_hold_amt"] = "{:,.2f}".format(float(d["intragov_hold_amt"]))
-        
+
         d["tot_pub_debt_out_amt"] = "{:,.2f}".format(float(d["tot_pub_debt_out_amt"]))
 
         if last_year is not None:
@@ -125,55 +124,3 @@ async def public_debt(request: Request):
     )
 
 
-async def get_public_debt():
-    dates_list = [
-        "2000-01-03",
-        "2001-01-05",
-        "2002-01-03",
-        "2003-01-03",
-        "2004-01-05",
-        "2005-01-03",
-        "2006-01-03",
-        "2007-01-03",
-        "2008-01-03",
-        "2009-01-06",
-        "2010-01-05",
-        "2011-01-03",
-        "2012-01-03",
-        "2013-01-03",
-        "2014-01-02",
-        "2015-01-02",
-        "2016-01-04",
-        "2017-01-03",
-        "2018-01-02",
-        "2019-01-02",
-        "2020-01-02",
-        "2021-01-04",
-        "2022-01-03",
-        "2023-01-03",
-        "2024-01-02",
-    ]
-
-    current_date = datetime.datetime.now()
-    days_to_subtract = (current_date.weekday() - 2) % 7 + 7
-    last_wednesday = current_date - datetime.timedelta(days=days_to_subtract)
-    last_wednesday_str = last_wednesday.strftime("%Y-%m-%d")
-    dates_list.append(last_wednesday_str)
-    dates_list.sort()
-    t0 = time.time()
-    tasks = [
-        debt_api_call(debt_date=d) for d in tqdm(dates_list, ascii=False, leave=True)
-    ]
-    debt_list = [task.result() for task in tasks]
-    print(f"Debt API call time taken: {time.time() - t0:.3f}")
-    return sorted(debt_list, key=lambda d: d["record_date"])
-
-
-@unsync
-async def debt_api_call(debt_date: str):
-
-    url = f"https://api.fiscaldata.treasury.gov/services/api/fiscal_service/v2/accounting/od/debt_to_penny?filter=record_date:eq:{debt_date}"
-    response = await client.get(url)
-    resp = response.json()
-    logger.debug(f"Debt API call response: {resp} for date {debt_date}")
-    return resp["data"][0]
