+++ templates/base_navbar.html
@@ -12,7 +12,7 @@
       <a href="/pages/about" class="nav-link">About</a>
     </li>
     <li class="nav-item d-none d-sm-inline-block">
-      <a href="/pages/public-debt" class="nav-link">US Debt</a>
+      <a href="/pages/public-debt" class="nav-link">USDebt</a>
     </li>
     <!-- Search From Navbar -->
     <li class="nav-item d-none d-sm-inline-block">
