+++ templates/interesting-data.html
@@ -1,46 +1,24 @@
 {% extends "base.html" %}
 {% block page_title %}| About{% endblock %}
-{% block page_stylesheet %}{% endblock %}
+{% block page_stylesheet %}
+<!-- DataTables -->
+<link
+  href="https://cdn.datatables.net/v/bs5/jq-3.7.0/dt-2.0.5/b-3.0.2/b-html5-3.0.2/b-print-3.0.2/sc-2.4.1/sb-1.7.1/datatables.min.css"
+  rel="stylesheet">
+<!-- Stylesheet for DataTables Bootstrap 4 -->
+<link rel="stylesheet"
+  href="{{ url_for('static' , path='plugins/datatables-responsive/css/responsive.bootstrap4.min.css') }}">
+<!-- Stylesheet for DataTables Responsive Bootstrap 4 -->
+<link rel="stylesheet"
+  href="{{ url_for('static' , path='plugins/datatables-buttons/css/buttons.bootstrap4.min.css') }}">
+<!-- Stylesheet for DataTables Buttons Bootstrap 4 -->
+{% endblock %}
 {% block page_breadcrumb %} About{% endblock %}
 {% block page_head_title %}About Page{% endblock %}
 {% block body %}
 <!-- Main content -->
 <section class="content">
-  <div class="row">
-    <div class="col-12">
-      <table id="failTable" class="table table-striped" style="width:100%">
-        <thead>
-          <tr>
-            <th>record_date</th>
-            <th>debt_held_public_amt</th>
-            <th>intragov_hold_amt</th>
-            <th>tot_pub_debt_out_amt</th>
-            <th>debt_growth</th>
-          </tr>
-        </thead>
-        <tbody>
-          {% for data in debt %}
-          <tr>
-            <td>{{data.record_date}}</td>
-            <td>{{data.debt_held_public_amt}}</td>
-            <td>{{data.intragov_hold_amt}}</td>
-            <td>{{data.tot_pub_debt_out_amt}}</td>
-            <td>{{data.debt_growth}}</td>
-          </tr>
-          {% endfor %}
-        </tbody>
-        <tfoot>
-          <tr>
-            <th>record_date</th>
-            <th>debt_held_public_amt</th>
-            <th>intragov_hold_amt</th>
-            <th>tot_pub_debt_out_amt</th>
-            <th>debt_growth</th>
-          </tr>
-        </tfoot>
-      </table>
-    </div>
-  </div>
+{% include 'api/us-debt.html' %}
 </section>
 <!-- /.content-wrapper -->
 </div>
@@ -55,4 +33,27 @@
 {% endblock %}
 {% block page_scripts %}
 <!-- page scripts-->
+<script src="https://code.jquery.com/jquery-3.7.1.min.js"
+  integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"></script>
+
+<script src="https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.2.7/pdfmake.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.2.7/vfs_fonts.js"></script>
+<script
+  src="https://cdn.datatables.net/v/bs5/jq-3.7.0/dt-2.0.5/b-3.0.2/b-html5-3.0.2/b-print-3.0.2/sc-2.4.1/sb-1.7.1/datatables.min.js"></script>
+
+
+  <script>
+    $(function () {
+      $('#debtTable').DataTable({
+        "paging": true,
+        "lengthChange": true,
+        "searching": true,
+        "ordering": true,
+        "info": true,
+        "autoWidth": true,
+        dom: 'Bfrtip',
+        buttons: ['copy', 'csv','pdf']
+      });
+    });
+  </script>
 {% endblock %}
